Move IntoRangeBounds trait into traits.rs
